+++ .gitignore
@@ -38,3 +38,4 @@ out/
 
 /src/main/resources/application-oauth.properties
 /src/main/resources/application-database.properties
+/src/main/resources/application-mail.properties
