+++ src/main/java/kr/or/cola/backend/user/config/auth/SecurityConfig.java
@@ -1,0 +1,32 @@
+package kr.or.cola.backend.user.config.auth;
+
+import kr.or.cola.backend.user.config.auth.dto.CustomOAuth2UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@RequiredArgsConstructor
+@EnableWebSecurity
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+    private final CustomOAuth2UserService customOAuth2UserService;
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+            .cors()
+                .and()
+            .csrf().disable().headers().frameOptions().disable()
+                .and()
+            .authorizeRequests()
+                .antMatchers("/", "/css/**", "/images/**", "/js/**").permitAll()
+                .anyRequest().authenticated()
+                .and()
+            .logout()
+                .logoutSuccessUrl("/")
+                .and()
+            .oauth2Login()
+                .userInfoEndpoint()
+                    .userService(customOAuth2UserService);
+    }
+}
